🚚 Rename: 북마크 관련 request type 파일 이동 (#191)
