+++ src/types/index.ts
@@ -4,6 +4,6 @@ import type { InputProps } from '@/types/components/input'
 import type {
   recruitmentSearchParams,
   lectureSearchParams,
-} from '@/types/api-request-types'
+} from '@/types/api-request-types/bookmarked-request-types'
 
 export type { Toast, InputProps, recruitmentSearchParams, lectureSearchParams }
